requirements.txt
streamlit>=1.32.0
google-generativeai>=0.5.0
langchain>=0.1.16
langchain-google-genai>=1.0.1
langchain-community>=0.0.36
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
tiktoken>=0.6.0
